docdigest/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20

# Async
httpx==0.28.1
anyio==4.8.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pgvector==0.3.6
alembic==1.14.1

# Task queue
celery[redis]==5.4.0
redis==5.2.1
flower==2.0.1

# LLM
anthropic==0.42.0

# Document parsing
PyPDF2==3.0.1
pdfplumber==0.11.4
python-docx==1.1.2
ebooklib==0.18
beautifulsoup4==4.12.3
lxml==5.3.0

# Text processing
tiktoken==0.8.0
nltk==3.9.1

# Embeddings
voyageai==0.3.2

# OCR (optional — requires system Tesseract)
pytesseract==0.3.13
Pillow==11.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0

# Dev utilities
rich==13.9.4
